test: fold locking tests into the segment store test file; drop "slice"
All SQLAlchemy segment store tests live in one file. The separate
locking module existed so the same tests could import against the
pre-overhaul partitioned store for the lock-ablation matrix; that
verification is done and recorded, so the split's constraint is spent.
The per-lock coverage map moves to a section comment.

Also replaces the "one slice per call" purge wording, which was
circular (a slice being defined as whatever one call does), with the
actual contract: each call reclaims at most max_segments segments --
in this store, one transaction that commits that progress or nothing --
and None means the store's default bound (_DEFAULT_PURGE_MAX_SEGMENTS,
renamed from _PURGE_SLICE_SEGMENTS).
